+++ src/main/java/com/freenow/user/blog/dto/Post.java
@@ -1,19 +1,19 @@
 package com.freenow.user.blog.dto;
 
 public class Post {
-    private float userId;
-    private float id;
+    private int userId;
+    private int id;
     private String title;
     private String body;
 
 
     // Getter Methods
 
-    public float getUserId() {
+    public int getUserId() {
         return userId;
     }
 
-    public float getId() {
+    public int getId() {
         return id;
     }
 
@@ -27,11 +27,11 @@ public class Post {
 
     // Setter Methods
 
-    public void setUserId(float userId) {
+    public void setUserId(int userId) {
         this.userId = userId;
     }
 
-    public void setId(float id) {
+    public void setId(int id) {
         this.id = id;
     }
